+++ src/main/java/net/silentchaos512/gear/api/material/IMaterialInstance.java
@@ -4,6 +4,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.util.ResourceLocation;
 import net.silentchaos512.gear.api.parts.MaterialGrade;
+import net.silentchaos512.gear.api.parts.PartType;
 
 import javax.annotation.Nullable;
 
@@ -18,4 +19,10 @@ public interface IMaterialInstance {
     ItemStack getItem();
 
     CompoundNBT write(CompoundNBT nbt);
+
+    int getColor(PartType partType, ItemStack gear);
+
+    default int getColor(PartType partType) {
+        return getColor(partType, ItemStack.EMPTY);
+    }
 }