+++ src/main/java/net/silentchaos512/gear/gear/material/MaterialInstance.java
@@ -123,4 +123,9 @@ public class MaterialInstance implements IMaterialInstance {
         nbt.put("Item", item.write(new CompoundNBT()));
         return nbt;
     }
+
+    @Override
+    public int getColor(PartType partType, ItemStack gear) {
+        return material.getColor(gear, partType);
+    }
 }